VS2013/15 aggregate changes of Intel and Microsoft
Status: Compiles in VS2015 and runs tests

List of TODOs:

  TODO: Fix tinydtls tick and locking issues.
  TODO: Fix disabled tests (PersistentStorageHandlerTest, usleep)
  TODO: Enable ocpmapi.dll shared lib (currently it's static)
  TODO: Remove hacks to solve error C2059: syntax error: 'constant'
          see: resource\include\StringConstants.h
  TODO: Address inconsistent style! Perhaps file a JIRA issue; CI
        should reject such issues before they enter the code base.
  TODO: Understand priority around enabling services.

Merger changelog:

* Adopted libtimer\timer.c changes from Microsoft
* Adopted Boost SCons installation changes from Intel
* Combined build_common\windows\SConscript from both Intel and
  Microsoft
* Adopted tinydtls code changes from Microsoft (INLINE_API macro,
  features wrapped in #ifdefs, etc)
* Adopted tinydtls SCons changes from Intel, tinydtls random seed
  generation from Intel
* Adopted ocrandom.c "QueryPerformanceCounter" from Microsoft.
* Adopted resource\examples from Intel
* Adopted liboc_logger and liboc from Intel
* Adopted resource/src from Intel
* Adopted boostification from Microsoft (BOOST_NOEXCEPT, etc.)
* Adopted "getopt" from Microsoft (multiple locations)
  - resource\csdk\stack\samples\linux\secure\occlientbasicops.cpp
  - resource\csdk\stack\samples\linux\SimpleClientServer\common.cpp
* Adopted SimpleClientServer changes from Microsoft
  - resource\csdk\stack\samples\linux\SimpleClientServer\
* Adopted oicgroup.c from Intel
* Adopted OC_EXPORT changes (and OC_EXPORT_DLL define macro)
  from Microsoft

Reviewed-on: https://gerrit.iotivity.org/gerrit/7627
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>

+++ resource/csdk/stack/src/ocpayloadconvert.c
@@ -65,7 +65,7 @@ OCStackResult OCConvertPayload(OCPayload* payload, uint8_t** outPayload, size_t*
     // TinyCbor Version 47a78569c0 or better on master is required for the re-allocation
     // strategy to work.  If you receive the following assertion error, please do a git-pull
     // from the extlibs/tinycbor/tinycbor directory
-    #define CborNeedsUpdating  (CborErrorOutOfMemory < CborErrorDataTooLarge)
+    #define CborNeedsUpdating  (((unsigned int)CborErrorOutOfMemory) < ((unsigned int)CborErrorDataTooLarge))
     OC_STATIC_ASSERT(!CborNeedsUpdating, "tinycbor needs to be updated to at least 47a78569c0");
     #undef CborNeedsUpdating
